refactor: split workflow_graph_source_ast.py under the 300-line cap (issue #275)
Completes the #258 asyncio-drain fix by resolving the follow-on file-size
debt it surfaced, instead of deferring it to a new issue (boy-scout rule,
§14 — the size violation is in the exact file this PR's own diff touches).

Splits the single 1037-line module into four cohesive concerns:
  * ap_sync_loop.py — the cross-loop sync/drain primitive (_SyncLoop),
    including the #258 fix (_drain_pending_tasks). Deletes _run(), a
    legacy sync wrapper with zero callers anywhere in the tree (dead
    code, §9).
  * workflow_graph_ast_symbols.py — AST symbol loading (label set,
    symbol-type mapping, per-label batched query construction).
  * workflow_graph_ast_edges.py — AST edge loading (~89 rel-table
    batched queries).
  * workflow_graph_ast_response.py — as_list, the shared AP response
    normalizer.
workflow_graph_source_ast.py becomes the ~300-line composition point;
WorkflowGraphASTSource's _symbol_batches_async/_edge_batches_async/
_verify_symbols_async stay thin instance-method delegates (not free
functions) because existing tests monkeypatch them per-instance.

Test-file updates were required, not optional: _SyncLoop's logger and
_SHUTDOWN_DRAIN_TIMEOUT_S now live in ap_sync_loop's own module globals,
so a test that monkeypatches/asserts against them has to target that
module directly — patching the workflow_graph_source_ast re-export is a
different attribute the drain code's bare-name lookup never reads.

Widening the scoped mutation run from the two originally-changed lines to
the whole (now-relocated) file surfaced 21 real survivors pre-existing in
_SyncLoop.__init__/_ensure_loop/run/run_iter that the #258 PR's narrower
scope never mutation-tested. Added 3 tests (thread is None before first
use, worker thread is named+daemonized, run_iter's stop-sentinel doesn't
collide with a legitimately-yielded None) and tightened 2 existing tests
to assert exact exception-message text. Down to 3 survivors, all
equivalent mutants with a source-backed rationale (2 carried over from
the original file, 1 newly documented and empirically verified against
the full infrastructure suite).

docs/module-inventory.md's infrastructure/ section lists all 5 modules
(criterion 3); public import paths (WorkflowGraphASTSource, _SyncLoop)
stay stable via re-export (criterion 4); doc-claims test count resynced
6557 -> 6560 (+3 new tests) via check_doc_claims.py/generate_repo_badges.py.

Fixes cdeust/Cortex#275
Refs cdeust/Cortex#258 (drain fix itself already merged on this branch)
